# api/openapi.yaml
openapi: 3.0.3
info:
  title: OurSay Account API
  description: Account registration & authentication. Email OTP is
    bootstrap/recovery only; WebAuthn passkeys are the day-to-day login. These
    passkeys are the ACCOUNT-LOGIN factor and are separate from the civic
    thread-signing keys in @oursay/public-record.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token
    cookieAuth:
      type: apiKey
      in: cookie
      name: oursay_session
...
paths:
...
  /v1/civic/threads/join:
    post:
      summary: "Join a thread: register this device's signer under the stable thread
        persona Pₜ"
      tags:
        - civic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                threadId:
                  type: string
                  description: Root entity id the thread is scoped to.
                jurisdiction:
                  type: string
                  description: Jurisdiction id, e.g. ab-ca-gov.
                signerPubkey:
                  type: string
                  description: This device's per-thread WebAuthn passkey pubkey (envelope
                    signerPubkey), SEC1 P-256 hex. The server is the authority
                    on Pₜ — first device wins, subsequent devices receive the
                    same Pₜ in the response.
                commitment:
                  type: string
                  description: Opaque binding commitment (sha256 hex); opening stays client-side.
                    Must match the previously-bound commitment on a
                    returning-device join.
              required:
                - threadId
                - jurisdiction
                - signerPubkey
                - commitment
              additionalProperties: false
      security: *a1
      responses:
        "200":
          description: The canonical thread persona Pₜ for the caller's (user, thread).
            Always returned by join.
          content:
            application/json:
              schema:
                type: object
                description: The canonical thread persona Pₜ for the caller's (user, thread).
                  Always returned by join.
                properties:
                  personaPubkey:
                    type: string
                    description: Stable thread persona pubkey Pₜ (envelope authorPubkey) — first
                      device's signer wins; subsequent devices receive the same
                      Pₜ.
                  personaName:
                    type: string
                    description: Globally-unique display name minted for Pₜ — what out-of-scope
                      viewers see in this thread.
                required:
                  - personaPubkey
                  - personaName
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/civic/appends/prepare:
    post:
      summary: "Prepare an append: server-derived fields for a civic intent"
      tags:
        - civic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                  description: Thread persona pubkey the action is authored as (must be the
                    caller's).
                intent:
                  type: object
                  description: A civic intent. `content` is the raw payload (kept private); the
                    server returns the fields to sign.
                  properties:
                    op:
                      type: string
                      enum:
                        - create
                        - update
                        - delete
                    type:
                      type: string
                      enum:
                        - post
                        - comment
                        - reaction
                        - petition
                        - petition_signature
                        - poll
                        - vote
                        - result
                    entityId:
                      type: string
                      description: Caller-chosen entity id (client and server must agree).
                    parent:
                      type: object
                      properties:
                        type:
                          type: string
                          enum:
                            - post
                            - comment
                            - reaction
                            - petition
                            - petition_signature
                            - poll
                            - vote
                            - result
                        id:
                          type: string
                      required:
                        - type
                        - id
                      additionalProperties: false
                    content:
                      description: Opaque payload (any JSON). Never published in plaintext.
                  required:
                    - op
                    - type
                    - entityId
                  additionalProperties: false
              required:
                - author
                - intent
              additionalProperties: false
      security: *a1
      responses:
        "200":
          description: Server-derived fields the client must sign over (mirrors
            RecordService.prepareAppend).
          content:
            application/json:
              schema:
                type: object
                description: Server-derived fields the client must sign over (mirrors
                  RecordService.prepareAppend).
                properties:
                  prevHash:
                    type: string
                    nullable: true
                  parentType:
                    type: string
                  parentId:
                    type: string
                  parentRevisionHash:
                    type: string
                  parentRevisionTxId:
                    type: string
                  rootEntityId:
                    type: string
                  nullifierParentId:
                    type: string
                  nullifier:
                    type: string
                required:
                  - prevHash
                  - rootEntityId
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
  /v1/civic/appends/submit:
    post:
      summary: Submit a client-signed WebAuthn (webauthn-es256) envelope into the
        verified record pool
      tags:
        - civic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                envelope:
                  type: object
                  additionalProperties: true
                salt:
                  type: string
                  description: Per-record content-commitment salt (random uuid/hex).
                content:
                  description: The civic payload this envelope commits to (any JSON).
              required:
                - envelope
                - salt
                - content
              additionalProperties: false
      security: *a1
      responses:
        "201":
          description: A reference to the appended (pooled) transaction.
          content:
            application/json:
              schema:
                type: object
                description: A reference to the appended (pooled) transaction.
                properties:
                  txId:
                    type: string
                  entityId:
                    type: string
                  txHash:
                    type: string
                required:
                  - txId
                  - entityId
                  - txHash
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details: {}
                    required:
                      - code
                      - message
                required:
                  - error
...